data/mt_gender/requirements.txt
google-cloud
docopt
gcloud
google-cloud-translate
# gcloud cli https://cloud.google.com/sdk/docs/install#linux